DUNE/Tasks: added inf(), war(), err(), cri(), debug() and trace() to the AbstractTask signature.
This is necessary to fix a circular class dependency between Tasks::Task and Entities::BasicEntity. The Task class is composed of several BasicEntities, while the BasicEntity and derived classes need the Task's printing mechanism for output.

Note that this is a stopgap measure, as the printing functions in reallity use the Task's "main entity" for logging purposes. Secondary entities should also be able to do so, but they still can't. A conceptually cleaner approach would be simply to let the "main entity" do the printing and logging, while the existing Task functions just call those, for backwards compatibility. Unfortunately it is very tricky to do so with varadic functions, hence the stopgap measure.