DUNE/Math/Matrix: Turn skew(double[]) into non-class function
The way skew(double data[3]) was defined, it was impossible to call
because it was a friend function that did not take it's friendly
class as an argument, so ADL cannot find it. Since skew(double[]) does
not need access to private members of class Matrix anyway, it can just
be made into a regular free function outside of the class, there is no
need for friend.