Fixed segfault in Seatrac Task because ticket pointer is not initialized
Fixed segfault in Seatrac Task because ticket pointer is not initialized.

The src/Trasnports/Seatrac/Task.cpp has a member variable Ticket* m_ticket.
However, that point is not initialized to NULL. This causes a segfault on
some systems in sendTxStatus(...) because the ticket is used without being
initialized. Normally, the check for NULL in clearTicket(...) would have
prevented this from happening.

Fixes #189.