Control/UAV/Ardupilot: Ensure hae_offset is not reset upon losing GPS fix
Previously, m_hae_offset would be reset to zero if a single GpsFix message indicated poor Gps conditions, and since it did not reset the m_offset_st flag, it would remain zero forever, causing the "WGS84 height" in DUNE to be equal to MSL. Now, the offset is only set to zero if the conversion flag is false, in which case m_offset_st is also reset to false. If Gps is lost, the old value of the offset is used.