Navigation/AUV/Navigation: initialize pointer to fix undefined behavior
Fixed initialization of `m_speed_model` pointer. Function
`onResourceRelease` is called after construction so an uninitialized
pointer was being deleted.

According to the standard (5.3.5 [expr.delete] / 2):

```
In the first alternative (delete object), the value of the operand
of delete may be a null pointer value, a pointer to a non-array
object created by a previous new-expression, or a pointer to a
subobject ([intro.object]) representing a base class of such an
object (Clause [class.derived]). If not, the behavior is undefined.
```